Significantly simply the inference demo
