+++ inference/mul_ops.py
@@ -9,7 +9,6 @@ CUDA_KERNEL = load(
     sources=[os.path.join(CUDA_FOLDER, "spqr_torch_lib.cpp"),
              os.path.join(CUDA_FOLDER, "spqr_cuda_kernel.cu")],
     extra_cuda_cflags=['-O3', '-arch=native']
-
 )
 
 torch.library.define(
