+++ inference_demo.py
@@ -1,13 +1,13 @@
-import os
 import sys
 import time
+from enum import IntEnum
+from typing import Tuple
 
+import numpy as np
 import torch
-import torch.nn as nn
-from tqdm import trange
+from transformers import AutoModelForCausalLM, LlamaTokenizer, AutoConfig, StaticCache
 
-from inference.module import LLama, Mode
-from modelutils import get_lm_logits
+from modelutils import suspend_nn_inits
 
 torch.autograd.set_grad_enabled(False)
 
@@ -21,163 +21,152 @@ except ModuleNotFoundError:
     has_safetensors = False
 
 
-@torch.no_grad()
-def get_inps(model, data_iterable, nsamples, seqlen=4096):
-    dev = model.device
-    """mocks model launch to collect inputs to the first model layer"""
-    print("catching inputs from data", flush=True)
-
-    layers = model.model.layers
-
-    if isinstance(data_iterable, torch.Tensor):
-        def batch_generator(testenc, seqlen, nsamples):
-            for i in range(nsamples):
-                batch = testenc[:, (i * seqlen): ((i + 1) * seqlen)].to(dev)
-                yield batch
-
-        data_iterable = batch_generator(data_iterable, seqlen, nsamples)
-
-    emb = model.get_input_embeddings()
-    emb_dev = emb.weight.device
-    if emb_dev.type != "cuda":
-        emb = emb.to(dev)
-        # opt has other embeddings
-        if model.config.model_type == "opt":
-            model.model.decoder.embed_positions = model.model.decoder.embed_positions.to(dev)
-            if hasattr(model.model.decoder, "project_in") and model.model.decoder.project_in:
-                model.model.decoder.project_in = model.model.decoder.project_in.to(dev)
-    dev = emb.weight.device  # now default device is the one where the embeddings are.
-    layer_dev = next(layers[0].parameters()).device
-    layers[0] = layers[0].to(dev)
-
-    dtype = next(iter(model.parameters())).dtype
-    print(f'nsamples: {nsamples} seqlen: {seqlen} hidden_size: {model.config.hidden_size}')
-    inps = torch.zeros((nsamples, seqlen, model.config.hidden_size), dtype=dtype, device=dev)
-
-    forward_arg_names = [
-        "attention_mask",
-        "position_ids",  # TODO: Remove
-    ]
-
-    cache = {"i": 0, "attention_mask": None, "alibi": None}
-
-    class Catcher(nn.Module):
-        def __init__(self, module):
-            super().__init__()
-            self.module = module
-
-        def forward(self, inp, **kwargs):
-            inps[cache["i"]] = inp
-            cache["i"] += 1
-            for forward_arg_name in forward_arg_names:
-                cache[forward_arg_name] = kwargs.get(forward_arg_name)
-            raise ValueError
-
-    layers[0] = Catcher(layers[0])
-    saved_num_threads = torch.get_num_threads()
-    torch.set_num_threads(min(16, saved_num_threads))
-    for batch in data_iterable:
-        try:
-            if isinstance(batch, (list, tuple)):
-                model(batch[0].to(dev))
-            elif isinstance(batch, torch.Tensor):
-                model(batch.to(dev))
-        except ValueError:
-            pass
-    torch.set_num_threads(saved_num_threads)
-    layers[0] = layers[0].module
-
-    layers[0] = layers[0].to(layer_dev)
-    model.get_input_embeddings().to(emb_dev)
-    print(f'config_type = {model.config.model_type}')
-    if model.config.model_type == "opt":
-        model.model.decoder.embed_positions = model.model.decoder.embed_positions.to(emb_dev)
-        if hasattr(model.model.decoder, "project_in") and model.model.decoder.project_in:
-            model.model.decoder.project_in = model.model.decoder.project_in.to(emb_dev)
-    torch.cuda.empty_cache()
-
-    forward_args = {k: cache[k] for k in forward_arg_names}
-    return inps, forward_args
+class Mode(IntEnum):
+    CUDA_PT = 0
+    CUDA_DENSE = 1
 
 
-@torch.no_grad()
-def perplexity_eval(model, testenc, randomize, tokenizer, dataset_name, nsamples):
-    seqlen = 4096
-    dev = model.device
-
-    inps, forward_args = get_inps(
-        model, testenc, nsamples, seqlen
-    )
-
-    if randomize:
-        inps = torch.randn_like(inps)
-
-    outs = torch.zeros_like(inps)
-    for k, v in forward_args.items():
-        forward_args[k] = v.to(dev) if isinstance(v, torch.Tensor) else v
-
-    layers = model.model.layers
-    for i in range(len(layers)):
-        layer = layers[i]
-        for j in range(nsamples):
-            outs[j] = layer(inps[j].unsqueeze(0), **forward_args)[0]
-        layers[i] = layer
-        inps, outs = outs, inps
-
-    for i in range(nsamples):
-        lm_logits = get_lm_logits(inps[i].to(dev), model)
-        print(tokenizer.batch_decode(lm_logits.argmax(-1), skip_special_tokens=True)[0])
-
-    testenc = testenc.to(dev)
-
-    nlls = []
-    for i in trange(nsamples):
-        lm_logits = get_lm_logits(inps[i].to(dev), model)
-        shift_logits = lm_logits[:, :-1, :].contiguous()
-        shift_labels = testenc[:, (i * seqlen): ((i + 1) * seqlen)][:, 1:]
-        loss_fct = nn.CrossEntropyLoss()
-        loss = loss_fct(shift_logits.view(-1, shift_logits.size(-1)), shift_labels.view(-1))
-        neg_log_likelihood = loss.float() * seqlen
-        nlls.append(neg_log_likelihood)
-    ppl = torch.exp(torch.stack(nlls).sum() / (nsamples * seqlen))
-    print(f"\n{dataset_name} perplexity = {ppl.item():.4f}\n")
+def decode_one_tokens(model, cur_token, input_pos, cache_position, past_key_values):
+    logits = model(
+        cur_token,
+        position_ids=input_pos,
+        cache_position=cache_position,
+        past_key_values=past_key_values,
+        return_dict=False,
+        use_cache=True
+    )[0]
+    new_token = torch.argmax(logits[:, -1], dim=-1)[:, None]
+    return new_token
 
 
-def old():
-    pretrained_model_path = sys.argv[1]
-    # uncompressed_model_path = sys.argv[2]
-    compressed_model_path = sys.argv[2]
-    mode = sys.argv[3]
-    # calibration_dataset = sys.argv[4]
-    # dev = sys.argv[5]
+class LLama:
+    def __init__(self, pretrained_model_path: str, quantized_model_path, flag, device='cuda', torchscript=False,
+                 backend=None):
+        self.flag = flag
+        self.device = device
+        self.dtype = torch.float16
+        self.torchscript = torchscript
 
-    m = Mode(int(mode))
+        self.config = AutoConfig.from_pretrained(pretrained_model_path, torchscript=self.torchscript,
+                                                 return_dict=True)
+        self.config.max_position_embeddings = 4096
 
-    with torch.no_grad():
-        model = LLama(pretrained_model_path, compressed_model_path, m)
-        text = 'The recipe for banana bread is '  # input()
-        s = time.time();
-        generated_text = model.generate(text, max_new_tokens=45);
-        e = time.time();
-        print(f'{generated_text}');
-        print(f'Duration = {e - s}s')
+        if flag == Mode.CUDA_PT:
+            with suspend_nn_inits():
+                with torch.no_grad():
+                    self.model = torch.load(quantized_model_path)
+        else:
+            with suspend_nn_inits():
+                with torch.no_grad():
+                    self.model = AutoModelForCausalLM.from_pretrained(
+                        pretrained_model_name_or_path=pretrained_model_path,
+                        trust_remote_code=True,
+                        torch_dtype=torch.half,
+                        config=self.config
+                    )
+
+        if self.torchscript:
+            self.model = torch.jit.script(self.model)
+
+        self.model.eval()
+
+        self.device = device
+
+        if backend is None:
+            self.model = self.model.to(device=self.device, dtype=self.dtype)
+        else:
+            self.model = torch.compile(self.model.to(device=self.device, dtype=self.dtype), backend=backend)
+
+        self.tokenizer = LlamaTokenizer.from_pretrained(pretrained_model_path, use_fast=False,
+                                                        torchscript=self.torchscript)
+        self.tokenizer.pad_token = self.tokenizer.eos_token
+
+    def generate(self, input_str, max_new_tokens) -> Tuple:
+        inputs = self.tokenizer(input_str, return_tensors="pt").to(device=self.device)
+
+        input_ids = inputs.input_ids
+        seq_len = input_ids.shape[1]
+
+        cache_position = torch.arange(seq_len, dtype=torch.int64, device=self.device)
+        generated_ids = torch.zeros(1, seq_len + max_new_tokens * 2, dtype=torch.int, device=self.device)
+        generated_ids[:, cache_position] = input_ids.to(self.device).to(torch.int)
+
+        past_key_values = StaticCache(self.config,
+                                      1,
+                                      seq_len + max_new_tokens * 2 + 1,
+                                      device=self.device,
+                                      dtype=torch.float16)
+
+        logits = self.model(
+            input_ids, cache_position=cache_position, past_key_values=past_key_values, return_dict=False, use_cache=True
+        )[0]
+        next_token = torch.argmax(logits[:, [-1]], dim=-1).to(torch.int)
+        generated_ids[:, [seq_len]] = next_token
+
+        torch._dynamo.config.capture_scalar_outputs = True
+
+        forward_time_s = []
+        with torch.no_grad():
+            # Compile the CUDA graph
+            decode_one_tokens_compiled = torch.compile(decode_one_tokens, mode='reduce-overhead', fullgraph=False)
+
+            # Generate tokens one by one
+            cache_position = torch.tensor([seq_len + 1], device="cuda")
+            for _ in range(1, max_new_tokens):
+                # with torch.nn.attention.sdpa_kernel([SDPBackend.MATH]):
+                with torch.backends.cuda.sdp_kernel(enable_flash=False, enable_mem_efficient=False, enable_math=True):
+                    start_time = time.time()
+                    next_token = decode_one_tokens_compiled(self.model, next_token.clone(), None, cache_position,
+                                                            past_key_values)
+                    generated_ids[:, cache_position] = next_token.int()
+                    end_time = time.time()
+                    print(f'duration = {end_time - start_time}')
+                    forward_time_s.append(end_time - start_time)
+
+                cache_position += 1
+                # print(self.tokenizer.decode(generated_ids[0]))
+
+        return self.tokenizer.decode(generated_ids[0]), forward_time_s
 
 
 if __name__ == "__main__":
-    pretrained_model_path = sys.argv[1]
-    # uncompressed_model_path = sys.argv[2]
-    compressed_model_path = sys.argv[2]
-    mode = sys.argv[3]
-    # calibration_dataset = sys.argv[4]
-    # dev = sys.argv[5]
+    import argparse
 
-    m = Mode(int(mode))
+    parser = argparse.ArgumentParser(add_help=True)
+
+    parser.add_argument(
+        "pretrained_model_path",
+        type=str,
+        help="Path to the model to the pretrained model",
+    )
+    parser.add_argument(
+        "compressed_model_path",
+        type=str,
+        help="Path to the compressed .pt model",
+    )
+    parser.add_argument(
+        "execution_mode",
+        type=int,
+        choices=[0, 1],
+        help="If set to 0, will evaluate the dense pretrained model. "
+             "If set to 1, will evaluate the spqr-quantized model",
+    )
+
+    args = parser.parse_args()
+
+    m = Mode(args.execution_mode)
 
     with torch.no_grad():
-        model = LLama(pretrained_model_path, compressed_model_path, m)
+        model = LLama(args.pretrained_model_path, args.compressed_model_path, m)
         text = 'The recipe for banana bread is '  # input()
-        s = time.time();
-        generated_text = model.generate(text, max_new_tokens=128);
-        e = time.time();
-        print(f'{generated_text}');
-        print(f'Duration = {e - s}s')
+        s = time.time()
+        generated_text, timings_s = model.generate(text, max_new_tokens=128)
+        e = time.time()
+        print(f'{generated_text}')
+
+        print(f'Total duration = {e - s}s')
+
+        durations = np.array(timings_s[5:])
+
+        print(f'Mean duration after caching initial input = {durations.mean()}')
+        print(f'Median duration after caching initial input = {np.median(durations)}')
